ecom-microservices/payment-service/requirements.txt
django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
requests==2.31.0
django-cors-headers==4.3.1
